requirements.txt
streamlit
llama-index-core
llama-index-llms-groq
llama-index-embeddings-huggingface
llama-index-readers-file
nest_asyncio
python-dotenv